+++ databento/live/session.py
@@ -6,7 +6,6 @@ import logging
 import queue
 import threading
 from collections.abc import Iterable
-from numbers import Number
 from typing import IO, Callable, Final
 
 import databento_dbn
@@ -335,7 +334,7 @@ class Session:
         self,
         dataset: Dataset | str,
         schema: Schema | str,
-        symbols: Iterable[str] | Iterable[Number] | str | Number = ALL_SYMBOLS,
+        symbols: Iterable[str | int] | str | int = ALL_SYMBOLS,
         stype_in: SType | str = SType.RAW_SYMBOL,
         start: str | int | None = None,
     ) -> None:
@@ -345,11 +344,11 @@ class Session:
 
         Parameters
         ----------
-        dataset : Dataset, str
+        dataset : Dataset or str
             The dataset for the subscription.
         schema : Schema or str
             The schema to subscribe to.
-        symbols : Iterable[str | Number] or str or Number, default 'ALL_SYMBOLS'
+        symbols : Iterable[str | int] or str or int, default 'ALL_SYMBOLS'
             The symbols to subscribe to.
         stype_in : SType or str, default 'raw_symbol'
             The input symbology type to resolve from.
@@ -503,7 +502,7 @@ class Session:
     async def _subscribe_task(
         self,
         schema: Schema | str,
-        symbols: Iterable[str] | Iterable[Number] | str | Number = ALL_SYMBOLS,
+        symbols: Iterable[str | int] | str | int = ALL_SYMBOLS,
         stype_in: SType | str = SType.RAW_SYMBOL,
         start: str | int | None = None,
     ) -> None: